+++ classes/class.ocpcRouterGUI.php
@@ -1,8 +1,5 @@
 <?php
 
-require_once __DIR__ . "/../vendor/autoload.php";
-require_once __DIR__ . '/../../../../Repository/RepositoryObject/OpenCast/vendor/autoload.php';
-
 use ILIAS\DI\Container;
 use srag\DIC\OpenCast\Exception\DICException;
 use srag\DIC\OpencastPageComponent\DICTrait;
@@ -26,6 +23,7 @@ use srag\Plugins\Opencast\UI\Input\Plupload;
 use srag\Plugins\Opencast\DI\OpencastDIC;
 use srag\Plugins\Opencast\Util\FileTransfer\UploadStorageService;
 use srag\Plugins\OpencastPageComponent\Authorization\TokenRepository;
+use srag\Plugins\Opencast\Container\Init;
 
 /**
  * Class ocpcRouterGUI
@@ -55,10 +53,7 @@ class ocpcRouterGUI
      * @var ACLUtils
      */
     private $acl_utils;
-    /**
-     * @var \srag\Plugins\Opencast\DI\OpencastDIC
-     */
-    private $opencast_dic;
+    private OpencastDIC $legacy_container;
     /**
      * @var Container
      */
@@ -67,42 +62,43 @@ class ocpcRouterGUI
      * @var \ilGlobalTemplateInterface
      */
     private $main_tpl;
-    /**
-     * @var \ilOpencastPageComponentPlugin
-     */
-    private $plugin;
+    private \ilOpencastPageComponentPlugin $plugin;
     private ilOpenCastPlugin $opencast_plugin;
+
+    private \srag\Plugins\Opencast\Container\Container $container;
 
     public function __construct()
     {
         global $DIC;
-        global $DIC, $opencastContainer;
-        $this->plugin = ilOpencastPageComponentPlugin::getInstance();
-        $this->main_tpl = $DIC->ui()->mainTemplate();
         $this->dic = $DIC;
-        PluginConfig::setApiSettings();
-        $this->opencast_dic = OpencastDIC::getInstance();
-        $this->opencast_plugin =$this->opencast_dic->plugin();
-        $this->opencast_dic->overwriteService(
+        $this->plugin = ilOpencastPageComponentPlugin::getInstance();
+        $this->container = Init::init($DIC);
+        $this->main_tpl = $DIC->ui()->mainTemplate();
+
+        $this->opencast_plugin = $this->container->plugin();
+        $this->legacy_container = $this->container->legacy();
+
+        $this->legacy_container->overwriteService(
             'upload_handler',
-            function (): \xoctFileUploadHandlerGUI {
-                return new xoctFileUploadHandlerGUI(
-                    $this->opencast_dic->upload_storage_service(),
-                    $this->dic->ctrl()->getLinkTargetByClass(
-                        [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class], 'upload'
-                    ),
-                    $this->dic->ctrl()->getLinkTargetByClass(
-                        [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class], 'info'
-                    ),
-                    $this->dic->ctrl()->getLinkTargetByClass(
-                        [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class], 'remove'
-                    )
-                );
-            }
+            fn (): \xoctFileUploadHandlerGUI => new xoctFileUploadHandlerGUI(
+                $this->legacy_container->upload_storage_service(),
+                $this->dic->ctrl()->getLinkTargetByClass(
+                    [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class],
+                    'upload'
+                ),
+                $this->dic->ctrl()->getLinkTargetByClass(
+                    [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class],
+                    'info'
+                ),
+                $this->dic->ctrl()->getLinkTargetByClass(
+                    [ilObjPluginDispatchGUI::class, ocpcRouterGUI::class, xoctFileUploadHandlerGUI::class],
+                    'remove'
+                )
+            )
         );
 
-        $this->event_repository = $opencastContainer[EventAPIRepository::class];
-        $this->series_repository = $opencastContainer->get(SeriesAPIRepository::class);
+        $this->event_repository = $this->container[EventAPIRepository::class];
+        $this->series_repository = $this->container[SeriesAPIRepository::class];
     }
 
     public function executeCommand(): void
@@ -132,8 +128,8 @@ class ocpcRouterGUI
                 }
                 $xoctPlayerGUI = new xoctPlayerGUI(
                     $this->event_repository,
-                    $this->opencast_dic->paella_config_storage_service(),
-                    $this->opencast_dic->paella_config_service_factory()
+                    $this->legacy_container->paella_config_storage_service(),
+                    $this->legacy_container->paella_config_service_factory()
                 );
                 $xoctPlayerGUI->streamVideo();
                 break;
@@ -168,7 +164,7 @@ class ocpcRouterGUI
 
         $with_terms_of_use = !ToUManager::hasAcceptedToU($this->dic->user()->getId());
 
-        $form = $this->opencast_dic->event_form_builder()->upload(
+        $form = $this->legacy_container->event_form_builder()->upload(
             $form_action_by_class,
             $with_terms_of_use
         );
@@ -183,15 +179,15 @@ class ocpcRouterGUI
                 // this is necessary because the 'required'-function of the checkbox doesn't work currently
                 // otherwise, $data would just be null
                 $this->main_tpl->setOnScreenMessage(
-                    'failure', self::plugin()->getPluginObject()->txt('event_error_alert_accpet_terms_of_use')
+                    'failure',
+                    self::plugin()->getPluginObject()->txt('event_error_alert_accpet_terms_of_use')
                 );
                 $this->dic->mainTemplate()->loadStandardTemplate();
                 $this->dic->mainTemplate()->setContent($this->dic->ui()->renderer()->render($form));
                 $this->dic->mainTemplate()->printToStdout();
                 return;
-            } else {
-                ToUManager::setToUAccepted($this->dic->user()->getId());
             }
+            ToUManager::setToUAccepted($this->dic->user()->getId());
         }
 
         if (!$data) {
@@ -216,7 +212,7 @@ class ocpcRouterGUI
             new UploadEventRequest(
                 new UploadEventRequestPayload(
                     $metadata,
-                    $this->opencast_dic->acl_utils()->getBaseACLForUser(xoctUser::getInstance($this->dic->user())),
+                    $this->legacy_container->acl_utils()->getBaseACLForUser(xoctUser::getInstance($this->dic->user())),
                     new Processing(
                         PluginConfig::getConfig(PluginConfig::F_WORKFLOW),
                         $this->getDefaultWorkflowParameters($data['workflow_configuration']['object'] ?? null)
@@ -225,7 +221,7 @@ class ocpcRouterGUI
                 )
             )
         );
-        $this->opencast_dic->upload_storage_service()->delete($data['file']['file']['id']);
+        $this->legacy_container->upload_storage_service()->delete($data['file']['file']['id']);
 
         $this->main_tpl->setOnScreenMessage('success', $this->plugin->txt('msg_created'), true);
         $this->cancel();
@@ -243,10 +239,13 @@ class ocpcRouterGUI
 
             // Here we only get the admin default workflow parameters.
             $default_value = $param->getDefaultValueAdmin();
-
-            if (!isset($from_data->$id) && $default_value == WorkflowParameter::VALUE_ALWAYS_ACTIVE) {
-                $default_parameter->$id = "true";
+            if (isset($from_data->$id)) {
+                continue;
             }
+            if ($default_value != WorkflowParameter::VALUE_ALWAYS_ACTIVE) {
+                continue;
+            }
+            $default_parameter->$id = "true";
         }
         return $default_parameter;
     }