+++ plugin.php
@@ -1,6 +1,7 @@
 <?php
+
 $id = "ocpc";
-$version = "8.0.1";
+$version = "8.1.0";
 $ilias_min_version = "8.0";
 $ilias_max_version = "8.999";
 $responsible = 'Universität Bern';